fixed routs

+++ Develop/routes/api/product-routes.js
@@ -1,15 +1,15 @@
-const router = require('express').Router();
-const { Product, Category, Tag, ProductTag } = require('../../models');
+const router = require("express").Router();
+const { Product, Category, Tag, ProductTag } = require("../../models");
 
 // The `/api/products` endpoint
 
 // get all products
-router.get('/',async(req, res) => {
+router.get("/", async (req, res) => {
   // find all products
   // be sure to include its associated Product and Tag data
   try {
     const productData = await Product.findAll({
-      include: [{ model: Category },{model:Tag, through:ProductTag}],
+      include: [{ model: Category }, { model: Tag, through: ProductTag }],
     });
     res.status(200).json(productData);
   } catch (err) {
@@ -18,13 +18,13 @@ router.get('/',async(req, res) => {
 });
 
 // get one product
-router.get('/:id',async(req, res) => {
+router.get("/:id", async (req, res) => {
   // find a single product by its `id`
   // be sure to include its associated Product and Tag data
   try {
-    const productData = await Product.findByPkasync(req.params.id, {
+    const productData = await Product.findByPk(req.params.id, {
       // JOIN with locations, using the Trip through table
-      include: [{ model: Category },{model:Tag, through:ProductTag}]
+      include: [{ model: Category }, { model: Tag, through: ProductTag }],
     });
 
     if (!productData) {
@@ -39,12 +39,11 @@ router.get('/:id',async(req, res) => {
 });
 
 // create new product
-router.post('/',async(req, res) => {
-
+router.post("/", async (req, res) => {
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      ifasync(req.body.tagIds.length) {
+      if (req.body.tagIds.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -64,7 +63,7 @@ router.post('/',async(req, res) => {
 });
 
 // update product
-router.put('/:id',async(req, res) => {
+router.put("/:id", async (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
@@ -72,27 +71,26 @@ router.put('/:id',async(req, res) => {
     },
   })
     .then((product) => {
-      ifasync(req.body.tagIds && req.body.tagIds.length) {
-        
+      if (req.body.tagIds && req.body.tagIds.length) {
         ProductTag.findAll({
-          where: { product_id: req.params.id }
+          where: { product_id: req.params.id },
         }).then((productTags) => {
           // create filtered list of new tag_ids
           const productTagIds = productTags.map(({ tag_id }) => tag_id);
           const newProductTags = req.body.tagIds
-          .filter((tag_id) => !productTagIds.includes(tag_id))
-          .map((tag_id) => {
-            return {
-              product_id: req.params.id,
-              tag_id,
-            };
-          });
+            .filter((tag_id) => !productTagIds.includes(tag_id))
+            .map((tag_id) => {
+              return {
+                product_id: req.params.id,
+                tag_id,
+              };
+            });
 
-            // figure out which ones to remove
+          // figure out which ones to remove
           const productTagsToRemove = productTags
-          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-          .map(({ id }) => id);
-                  // run both actions
+            .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+            .map(({ id }) => id);
+          // run both actions
           return Promise.all([
             ProductTag.destroy({ where: { id: productTagsToRemove } }),
             ProductTag.bulkCreate(newProductTags),
@@ -108,7 +106,7 @@ router.put('/:id',async(req, res) => {
     });
 });
 
-router.delete('/:id',async(req, res) => {
+router.delete("/:id", async (req, res) => {
   // delete one product by its `id` value
   try {
     const productData = await Product.destroy({
